requirements.txt
asn1crypto==0.24.0
autopep8==1.4.1
cachetools==2.1.0
cffi==1.11.5
config==0.3.9
confusable-homoglyphs==3.2.0
cryptography==2.3.1
dj-database-url==0.5.0
Django==1.11.15
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-contrib-comments==1.9.0
django-heroku==0.3.1
django-multiselectfield==0.1.8
django-registration==2.4.1
django-tinymce==2.7.0
djangorestframework==3.8.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
numpy==1.15.2
oauth2client==4.1.3
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.19
pyOpenSSL==18.0.0
python-decouple==3.1
python-google-calendar-api==0.5
pytz==2018.5
rsa==4.0
six==1.11.0
uritemplate==3.0.0
whitenoise==4.1
